@@ -2,23 +2,28 @@ name: meditation_app
 version: 1.0.0+1
 
 environment:
-  sdk: '>=2.18.1 <3.0.0'
+  sdk: '>=3.4.0 <4.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
   awesome_snackbar_content: ^0.1.3
   badges: ^3.1.2
-  cupertino_icons: ^1.0.2
+  cupertino_icons: ^1.0.8
   draggable_fab: ^0.1.4
-  flashy_tab_bar2: ^0.0.4
-  flutter_native_splash: ^2.2.10+1
+  flashy_tab_bar2: ^0.0.10
+  flutter_native_splash: ^2.4.1
   google_fonts: ^6.2.1
   lottie: ^3.1.2
-  syncfusion_flutter_calendar: ^26.1.42
-  audioplayers: ^5.2.1
+  syncfusion_flutter_calendar: ^26.2.7
+  audioplayers: ^6.0.0
   carousel_slider: ^4.2.1
   record: ^5.1.2
+  change_app_package_name: ^1.3.0
+  sqflite: ^2.3.3+1
+  path: ^1.9.0
+  path_provider: ^2.1.4
+  avatar_glow: ^3.0.1
 
 dev_dependencies:
   flutter_test:
@@ -41,16 +46,18 @@ flutter_icons:
 # flutter pub run flutter_native_splash:create
 flutter_native_splash:
   color: "#ffffff"
-  image: assets/flutter.png
+  image: assets/logo.png
   web: true
   android_12:
-    image: assets/flutter.png
+    image: assets/logo.png
     icon_background_color: "#ffffff"
 
 flutter:
   uses-material-design: true
   assets:
-    - assets/flutter.png
     - assets/logo.png
     - assets/calm.mp3
     - assets/meditation.json
+    - assets/fire.json
+    - assets/wind.json
+    - assets/ocean.json
