@@ -13,29 +13,4 @@ final font = GoogleFonts.montserrat(
   fontWeight: FontWeight.w500,
 );
 
-List jumlah = ["Mood Swing", "Stress", "Depression", "Healing", "Relax"];
-
-List lorem = [
-  "Deserunt culpa incididunt exercitation tempor do adipisicing. Adipisicing est labore enim consectetur amet aute ullamco quis sint mollit voluptate ad sunt laborum.",
-  "During loving kindness meditation, you focus benevolent and loving energy toward yourself and others.",
-  "Outdoor conentration meditation, the object is a flower.",
-  "5-minute intro to meditation. relax and inhale to start.",
-];
-
-List time = [
-  "45 Mins",
-  "20 Mins",
-  "30 Mins",
-  "5 Mins",
-];
-
-List hour = [
-  "12:00 AM",
-  "2:30 PM",
-  "3:45 PM",
-  "3:50 PM",
-  "4:44 PM",
-  "5:55 PM",
-];
-
-List name = ["James Madchen", "Kate London", "David Willson", "Tommy Adam"];
+List moodStates = ["Mood Swing", "Stress", "Depression", "Healing", "Relax"];
